refactor: consolidate channel count to hardware profile
Remove MULTIPLEXER_CHANNEL_COUNT and CHANNEL_COUNT from constants.h.
Static arrays use MAX_CHANNEL_COUNT (compile-time). Runtime loops use
globalHwProfile->muxChannelCount + 1.

Move readEfuseProvisioningData() from utils to hardware_profile — eFuse
data is hardware/PCB data and belongs with its owner. Remove utils.h
include from hardware_profile.h (circular dependency). Remove eFuse
fields from MQTT provisioning request (known at manufacturing time).
