+++ source/include/modbustcp.h
@@ -19,7 +19,7 @@
 #define START_REGISTERS_METER_VALUES 100 // Before this, data that is not related to energy values (like time)
 #define LOWER_LIMIT_CHANNEL_REGISTERS 1000
 #define STEP_CHANNEL_REGISTERS 100
-#define UPPER_LIMIT_CHANNEL_REGISTERS (LOWER_LIMIT_CHANNEL_REGISTERS + (CHANNEL_COUNT) * STEP_CHANNEL_REGISTERS)
+#define UPPER_LIMIT_CHANNEL_REGISTERS (LOWER_LIMIT_CHANNEL_REGISTERS + (MAX_CHANNEL_COUNT) * STEP_CHANNEL_REGISTERS)
 
 namespace ModbusTcp
 {
